### 2019-07-08 #24 @whedon
[PRE REVIEW]: HydroPowerModels.jl: A Julia/JuMP Package forHydrothermal Economic Dispatch Optimization
**Submitting author:** @andrewrosemberg (<a href="http://orcid.org/0000-0002-6880-8845">Andrew Rosemberg</a>)
**Repository:** <a href="https://github.com/andrewrosemberg/HydroPowerModels.jl" target ="_blank">https://github.com/andrewrosemberg/HydroPowerModels.jl</a>

Thanks for submitting your paper to JuliaCon @andrewrosemberg. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@andrewrosemberg if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
