### 2019-07-08 16733bc
fix bulg exemple reports
