+++ examples/HydroValleys/case3_5years.jl
@@ -32,7 +32,7 @@ seed = 1221
 #' ## Load Case Specifications
 
 #' Data
-data = HydroPowerModels.parse_folder(joinpath(WEAVE_ARGS[:testcases_dir],"case3"));
+alldata = HydroPowerModels.parse_folder(joinpath(WEAVE_ARGS[:testcases_dir],"case3"));
 
 #' Plot power grid graph
 if plot_bool == true
@@ -60,7 +60,9 @@ end_time = time() - start_time
 (SDDP.termination_status(m.policygraph), end_time)
 
 #' Bounds
-plot_bound(m)
+if plot_bool == true
+    plot_bound(m)
+end
 
 #' ## Simulation
 import Random
@@ -69,9 +71,11 @@ results = HydroPowerModels.simulate(m, 100);
 results
 
 #' ## Testing Results
-#' Bound
 using Test
-@test isapprox(sum(s[:stage_objective] for s in results[:simulations][1]), 9400.0, atol=1)
+#' Bound
+@test isapprox(SDDP.calculate_bound(m.policygraph), 76487.72, atol=1)
+#' Number of Simulations
+@test length(results[:simulations]) == 100
 
 #' ## Plot Aggregated Results
 if plot_bool == true
@@ -80,7 +84,7 @@ end
 
 #' # Annex 1: Case Summary
 if plot_bool == true
-    PowerModels.print_summary(data["powersystem"])
+    PowerModels.print_summary(alldata[1]["powersystem"])
 end
 
 #' # Annex 2: Plot Results
