+++ examples/HydroValleys/case3_ac.jl
@@ -33,7 +33,7 @@ seed = 1221
 #' ## Load Case Specifications
 
 #' Data
-data = HydroPowerModels.parse_folder(joinpath(WEAVE_ARGS[:testcases_dir],"case3"));
+alldata = HydroPowerModels.parse_folder(joinpath(WEAVE_ARGS[:testcases_dir],"case3"));
 
 #' Plot power grid graph
 if plot_bool == true
@@ -44,7 +44,7 @@ end
 params = create_param(  stages = 12, 
                         model_constructor_grid  = ACPPowerModel,
                         post_method             = PowerModels.post_opf,
-                        optimizer               = Ipopt.Optimizer)
+                        optimizer               = with_optimizer(Ipopt.Optimizer))
 
 #' ## Build Model
 #+ results =  "hidden"
@@ -60,13 +60,15 @@ end_time = time() - start_time
 (SDDP.termination_status(m.policygraph), end_time)
 
 #' Bounds
-plot_bound(m)
+if plot_bool == true
+    plot_bound(m)
+end
 
 #' ## Simulation
+#+ results =  "hidden"
 import Random
 Random.seed!(seed)
 results = HydroPowerModels.simulate(m, 100);
-results
 
 #' ## Results
 #' Objective
@@ -79,7 +81,7 @@ end
 
 #' # Annex 1: Case Summary
 if plot_bool == true
-    PowerModels.print_summary(data["powersystem"])
+    PowerModels.print_summary(alldata[1]["powersystem"])
 end
 
 #' # Annex 2: Plot Results
