+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#HydroPowerModels.jl-Documentation-1","page":"Home","title":"HydroPowerModels.jl Documentation","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"CurrentModule = HydroPowerModels","category":"page"},{"location":"#Overview-1","page":"Home","title":"Overview","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"HydroPowerModels.jl is a Julia/JuMP package for Hydrothermal Multistage Steady-State Power Network Optimization solved by Stochastic Dual Dynamic Programming (SDDP). ","category":"page"},{"location":"#","page":"Home","title":"Home","text":"Problem Specifications and Network Formulations are handled by PowerModels.jl. ","category":"page"},{"location":"#","page":"Home","title":"Home","text":"Solution method is handled by SDDP.jl.","category":"page"},{"location":"#Installation-1","page":"Home","title":"Installation","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"The current package is unregistered so you will need to add it as follows:","category":"page"},{"location":"#","page":"Home","title":"Home","text":"julia> ] add https://github.com/andrewrosemberg/HydroPowerModels.jl.git ","category":"page"},{"location":"getstarted/#Getting-started-1","page":"Manual","title":"Getting started","text":"","category":"section"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"Once PowerModels, SDDP and a solver (like Clp or Ipopt) are installed, and a case data folder (e.g. \"case3\") has been acquired, an Hydrothermal Multistage Steady-State Power Network Optimization can be executed.","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"First import the necessary packages:","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"using HydroPowerModels\nusing Ipopt, Clp","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"Load Case by passing the respective folder:","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"data = HydroPowerModels.parse_folder(\"case3_folderpath\")","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"Set Parameters to run, for example, an DC Economic Hydrothermal Dispatch:","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"params = create_param( stages = 12, \n                    model_constructor_grid  = DCPPowerModel,\n                    post_method             = PowerModels.post_opf,\n                    optimizer               = Clp.Optimizer);","category":"page"},{"location":"getstarted/#","page":"Manual","title":"Manual","text":"Build the Model and execute the SDDP method:","category":"page"},{"location":"getstarted/#","page":"Manual","title
[…]
