### 2019-07-09 #24 @andrewrosemberg
Here's suggestions for potential reviewers familiar with the subjects of optimal power flow and multistage stochastic programming:

- Jose Lara (@jd-lara)
- Bernardo da Costa (@bfpc)
- François Pacaud (@frapac)
- Luiz Carlos da Costa (@lccostajr)
