+++ Project.toml
@@ -12,15 +12,18 @@ Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
 GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
+Gurobi = "2e9cd046-0924-5485-92f1-d5272153d98b"
 InfrastructureModels = "2030c09a-7f63-5d83-885d-db604e0e9cc0"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+KNITRO = "67920dd8-b58e-52a8-8622-53c4cffbe346"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 MathOptFormat = "f4570300-c277-12e8-125c-4912f86ce65d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SCS = "c946c3f1-0d1f-5ce8-9dea-7daa1f7e2d13"
