### 2019-07-11 8b9e190
extend visualization to forward specs
