### 2019-07-15 #24 @vchuravy
@jd-lara, @bfpc, @farpac, @lccostajr would you be willing to review this paper?
