### 2019-07-15 #35 @whedon
[REVIEW]: HydroPowerModels.jl: A Julia/JuMP Package forHydrothermal Economic Dispatch Optimization
**Submitting author:** @andrewrosemberg (<a href="http://orcid.org/0000-0002-6880-8845">Andrew Rosemberg</a>)
**Repository:** <a href="https://github.com/andrewrosemberg/HydroPowerModels.jl" target ="_blank">https://github.com/andrewrosemberg/HydroPowerModels.jl</a>
