### 2019-07-15 #24 @jd-lara
Yes I can review the paper
