### 2019-07-15 andrewrosemberg/HydroPowerModels.jl#12 @frapac
Remove some dependencies in Project.toml?
Looking at the file [Project.toml](https://github.com/andrewrosemberg/HydroPowerModels.jl/blob/master/Project.toml), it occurs to me that we might want to reduce the dependencies of HydroPowerModels.jl.

For instance, do we need `Ipopt.jl`, `Weave.jl`, `Cairo.jl` or `GR.jl` in `Project.toml`? By looking at the code in the [module](https://github.com/andrewrosemberg/HydroPowerModels.jl/blob/master/src/HydroPowerModels.jl#L3-L4), it seems that we only need `PowerModels.jl`, `JuMP.jl`, `GLPK.jl` and `SDDP.jl` as main dependencies. Maybe we could move the other packages in `[extras]` for testing purposes?

Furthermore, do you think we could add `Plots.jl` as an optional dependency (for instance by exporting the plotting functions only if `Plots.jl` is already installed)? When deporting some computations in a cluster, we might want to load only the minimal set of packages IMO.

X-ref https://github.com/JuliaCon/proceedings-review/issues/35
