### 2019-07-15 #35 @frapac
@andrewrosemberg Do you mind if I open a PR on HydroPowerModels.jl to comment out your article?
