### 2019-07-15 #35 @andrewrosemberg
> @andrewrosemberg Do you mind if I open a PR on HydroPowerModels.jl to comment out your article?

I don't mind at all. Please feel free to open one!
