### 2019-07-15 250becb
create visualization folder and add soc stochastic case
