+++ Project.toml
@@ -8,6 +8,7 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
 GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
@@ -23,7 +24,6 @@ PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-SCS = "c946c3f1-0d1f-5ce8-9dea-7daa1f7e2d13"
 SDDP = "f4570300-c277-11e8-125c-4912f86ce65d"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
@@ -36,6 +36,8 @@ Plots = "~0.24.0"
 julia = "~1.0"
 
 [extras]
+Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
+ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
