### 2019-07-15 0f42cf1
rm codecov of visualization
