+++ .codecov.yml
@@ -1,1 +1,3 @@
 comment: false
+ignore:
+  - "./src/visualization"
