### 2019-07-15 andrewrosemberg/HydroPowerModels.jl#13 @andrewrosemberg
> The code coverage is low because of this last file. Do you think it could be relevant to move it a separate folder, dedicated to the plotting utilities (as src/visualization in the package SDDP.jl)?

Yes this is a good suggestion! Added in 250becbcb589eb0b2bf3ac71b6797bb714ca4ec1 .

> I would appreciate having a testing procedure that tests stochastic and non-linear examples as well. My opinion is that running a non-linear version of SDDP is one of the key ideas in HydroPowerModels.jl, and you might want to put more emphasis on it.

Case 3 is a small but stochastic case. Just added a SOC stochastic case in 250becbcb589eb0b2bf3ac71b6797bb714ca4ec1.
