+++ Project.toml
@@ -26,7 +26,6 @@ Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SDDP = "f4570300-c277-11e8-125c-4912f86ce65d"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TimeZones = "f269a46b-ccf7-5d73-abea-4c690281aa53"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
 
@@ -36,15 +35,16 @@ Plots = "~0.24.0"
 julia = "~1.0"
 
 [extras]
-Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
-ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
+ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
+GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
