+++ docs/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
