### 2019-07-16 andrewrosemberg/HydroPowerModels.jl#12 @andrewrosemberg
Good Point! Although, I must admit I'm not having much luck with the building of the weave reports in the documentation when I remove those visualization packages. Does anyone have a suggestion on how to organize the project tomls of the package? @Thuener @odow @joaquimg
