### 2019-07-16 andrewrosemberg/HydroPowerModels.jl#12 @andrewrosemberg
> You can add the solvers just for the tests, e.g.,
> https://github.com/odow/SDDP.jl/blob/0490bea2c46787e1d4d63a5491ea0106c7fe70cf/Project.toml#L24-L30
>
> I haven't tried optional dependencies.

Even if I run them also for the documentation?
