### 2019-07-16 a3cf772
first pass of the review
