+++ paper/paper.tex
@@ -17,6 +17,10 @@
 
 \begin{abstract}
 
+% <review @frapac>
+% at the end of the paragraph:
+% SDDP -> SDDP algorithm.
+% </review>
 \textit{HydroPowerModels.jl} is a Julia package for solving multistage, steady-state, hydro-dominated, power network optimization problems with stochastic dual dynamic programming. Our state-of-the-art open source tool is flexible enough for practitioners in the electrical sector to test new ideas in an efficient way. This tool was made possible by the Julia language and the surrounding ecosystem of packages. We use \textit{JuMP}, a package for mathematical programming modeling; \textit{PowerModels.jl}, a JuMP-extension for power network optimization; and \textit{SDDP.jl}, another extension that implements the SDDP.
 
 \end{abstract}
@@ -26,6 +30,11 @@
 The hydrothermal dispatch problem is important for the planning and operation of hydro-dominated electrical systems such as the Brazilian national grid. The objective is to coordinate generation, energy distribution, and hydro-storage management in order to minimize the cost of operation. In this context, the hydrothermal dispatch problem is a medium-term planning problem where uncertainties related to the hydrology (i.e., rainfall and other inflows) of the hydroelectric plants and consumer-demand have fundamental importance. Generation takes two main forms: i) hydro-generation with a low marginal cost, and ii) thermal-generation, with a high marginal cost. However, in hydro-dominated systems such as Brazil, there is often insufficient capacity of thermal generation to meet demand. Thus, the system-operator faces a trade-off between using water for cheap generation in the present, against conserving water for future periods of drought. Because of this trade-off, the hydrothermal dispatch problem is often modeled as a multistage stochastic problem \cite{pereira1991multi,maceira_ten_2008,terlaky_marginal_2017}.
  %\vskip 6pt
 
+% <review @frapac>
+% I recommend to cite "Bertsekas, Dynamic Programming and optimal control"  when first refering
+% to Stochastic Dynamic Programming. Citing the seminal article of Richard Bellman
+% seems also a good idea.
+% </review>
 Solving multistage stochastic programs, however, is a challenging numerical problem. The solution of the problem is intractable and common approximations such as stochastic dynamic programming can suffer from high dimensionality (frequently referred as the curse of dimensionality). One method that partially overcomes the curse of dimensionality is the stochastic dual dynamic programming algorithm (SDDP) of ~\cite{pereira1991multi}.
 
 A number of programs implementing SDDP are in-use around the world, ranging from unpublished implementations in academic institutions to professional software such as PSR's \textit{SDDP}\footnote{Confusingly, the ``SDDP'' acronym is used to denote a software product, the original algorithm of \cite{pereira1991multi}, and a more general class of algorithms inspired by the original SDDP algorithm.} \cite{psr_software_2019}. However, until recently, there was no fast, reliable, and open-source implementation of the SDDP algorithm. Without such a tool, researchers and practitioners have not had a common ground for the discussion and analysis of different hydrothermal dispatch formulations and their solutions.
@@ -39,7 +48,14 @@ The next sections are organized as follows: 1) A brief understanding of multista
 \section{Multistage Stochastic Optimization}
 \label{sec:multistochopt}
 
[…]
