+++ paper/header.tex
@@ -9,9 +9,15 @@
 \author[1, 3]{\\ Davi M. Valladão}
 \author[4]{Oscar Dowson}
 \affil[1]{Laboratory of Applied Mathematical Programming and Statistics (LAMPS)}
+% <review @frapac>
+% Eletrical -> Electrical
+% </review>
 \affil[2]{Department of Eletrical Engineering, PUC-RIO}
 \affil[3]{Department of Industrial Engineering, PUC-RIO}
 \affil[4]{Department of Industrial Engineering, Northwestern University, Evanston}
 
+% <review @frapac>
+% Stochastic -> Stochastic Programming
+% </review>
 \keywords{Julia, Optimization, Energy, Stochastic}
 
