+++ docs/src/apireference.md
@@ -1,0 +1,30 @@
+# API Reference
+
+## Hydro power model
+
+```@docs
+HydroPowerModels.hydrothermaloperation
+```
+
+## System data and parameters
+
+```@docs
+HydroPowerModels.parse_file_json
+HydroPowerModels.create_param
+```
+
+## Training the policy
+
+```@docs
+HydroPowerModels.train
+```
+
+## Visualize data and results
+
+```@docs
+HydroPowerModels.plotscenarios
+HydroPowerModels.plot_grid
+HydroPowerModels.plot_aggregated_results
+HydroPowerModels.plotresults
+HydroPowerModels.plot_bound
+```
