### 2019-07-16 andrewrosemberg/HydroPowerModels.jl#12 @joaquimg
Could the vizualization packages simply be testing dependencies first?
Later on we try to set them as optional.
