### 2019-07-16 a5b859e
move back test solvers
