+++ Project.toml
@@ -18,7 +18,6 @@ SDDP = "f4570300-c277-11e8-125c-4912f86ce65d"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TimeZones = "f269a46b-ccf7-5d73-abea-4c690281aa53"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
-ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 
 [compat]
 JuMP = "~0.19"
