### 2019-07-16 0aa2a26
rm unused solvers
