+++ Project.toml
@@ -22,6 +22,7 @@ Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 
 [compat]
 JuMP = "~0.19"
@@ -41,8 +42,9 @@ LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
+Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 
 [targets]
 test = ["Test","ECOS","GLPK"]
-docs = ["Cairo","Compose","ECOS","Fontconfig","GLPK","GR","GraphPlot","Ipopt","LightGraphs","Plots","Test","Weave"]
-src/visualization = ["Cairo","Compose","Fontconfig","GraphPlot","LightGraphs","Plots"]
+docs = ["Cairo","Compose","ECOS","Fontconfig","GLPK","GR","GraphPlot","Ipopt","LightGraphs","Plots","Test","Weave","Reel"]
+src/visualization = ["Cairo","Compose","Fontconfig","GraphPlot","LightGraphs","Plots","Reel"]
