+++ Project.toml
@@ -4,25 +4,25 @@ authors = ["andrewrosemberg <andrew@ele.puc-rio.br>"]
 version = "0.1.0"
 
 [deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
+Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
+GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 InfrastructureModels = "2030c09a-7f63-5d83-885d-db604e0e9cc0"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 MathOptFormat = "f4570300-c277-12e8-125c-4912f86ce65d"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SDDP = "f4570300-c277-11e8-125c-4912f86ce65d"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TimeZones = "f269a46b-ccf7-5d73-abea-4c690281aa53"
-Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
-Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
-Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
-GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
-LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 
 [compat]
 JuMP = "~0.19"
@@ -40,10 +40,10 @@ GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
-Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 
 [targets]
-test = ["Test","ECOS","GLPK"]
-docs = ["Cairo","Compose","ECOS","Fontconfig","GLPK","GR","GraphPlot","Ipopt","LightGraphs","Plots","Test","Weave","Reel"]
[…]
