### 2019-07-16 3916f4a
add weave to deps
