### 2019-07-16 andrewrosemberg/HydroPowerModels.jl#12 @andrewrosemberg
Fixed test packages to extra in 3916f4afcefb6f367a08b79303f096dde2390aae .
Although, it seams that, since the visualization functions are inside the module, I'm not able to move them to extra.
The only other package outside the scope of the module is Weave, but some how the target to "docs" folder is not being successful : https://travis-ci.com/andrewrosemberg/HydroPowerModels.jl/builds/119367814
