+++ src/HydroPowerModels.jl
@@ -77,14 +77,14 @@ function hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict)
         gatherusefulinfo!(data)
 
         # build eletric grid model using PowerModels
-        pm = PowerModels.build_generic_model(data["powersystem"], params["model_constructor_grid"], 
+        pm = PowerModels.build_model(data["powersystem"], params["model_constructor_grid"], 
             params["post_method"], jump_model=sp, setting = params["setting"])
         
         #if isforward # NOT YET IMPLEMENTED
-            #pm = PowerModels.build_generic_model(data["powersystem"], params["model_constructor_grid_forward"], 
+            #pm = PowerModels.build_model(data["powersystem"], params["model_constructor_grid_forward"], 
                 #params["post_method"], jump_model=sp, setting = params["setting"])
         #else
-            #pm = PowerModels.build_generic_model(data["powersystem"], params["model_constructor_grid_backward"], 
+            #pm = PowerModels.build_model(data["powersystem"], params["model_constructor_grid_backward"], 
             #    params["post_method"], jump_model=sp, setting = params["setting"])
         #end
         
