+++ Project.toml
@@ -6,11 +6,11 @@ version = "0.1.0"
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
+Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
-InfrastructureModels = "2030c09a-7f63-5d83-885d-db604e0e9cc0"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
@@ -46,5 +46,5 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
 
 [targets]
-docs = ["Cairo", "Compose", "ECOS", "Fontconfig", "GLPK", "GR", "GraphPlot", "Ipopt", "LightGraphs", "Plots", "Test", "Weave", "Reel"]
-test = ["Test", "ECOS", "GLPK"]
+docs = ["Cairo", "Compose", "Fontconfig", "GraphPlot", "LightGraphs", "Plots", "Weave", "Reel"]
+test = ["Test", "GLPK", "ECOS"]
