### 2019-07-16 b8213ff
add solvers to docs
