+++ Project.toml
@@ -46,5 +46,5 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
 
 [targets]
-docs = ["Cairo", "Compose", "Fontconfig", "GraphPlot", "LightGraphs", "Plots", "Weave", "Reel"]
+docs = ["Cairo", "Compose", "Fontconfig", "GraphPlot", "LightGraphs", "Plots", "Weave", "Reel","Test", "GLPK", "ECOS"]
 test = ["Test", "GLPK", "ECOS"]
