### 2019-07-17 7dc3246
add solvers to docs toml
