+++ docs/Project.toml
@@ -1,3 +1,7 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
+ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
