### 2019-07-17 a44660d
add manifest
