### 2019-07-17 #35 @Lukrosz
Hi @andrewrosemberg,

Here attached is a document with my reviews of your article ([HydroPowerModels.jl_article.pdf](https://github.com/JuliaCon/proceedings-review/files/3401264/HydroPowerModels.jl_article.pdf)). For now, they are mainly focused on the structure and grammatics of the article and details about the model. Later on, I will comment about the package.

Let me know when you have the chance to read the pdf.
