### 2019-07-21 #35 @frapac
Hi @andrewrosemberg,
please find [my final review](https://github.com/frapac/HydroPowerModels.jl/blob/fp/review/paper/review.md) of the paper. Let me know if you have any question about it!
