+++ paper/ref.bib
@@ -150,15 +150,17 @@ doi = {10.1137/15M1020575},
   year={2016}
 }
 
-@misc{sddpy,
-  title={sddpy Python for Stochastic Dual Dynamic Programming Algorithm},
-  year={2019},
-  url = {https://github.com/wanqiuchansheng/sddpy}
+@misc{Lingquan2019,
+  author = {Lingquan Ding, Shabbir Ahmed, Alexander Shapiro},
+  title={A Python package for multi-stage stochastic programming},
+  year = {2019},
+  month = may,
+  url = {http://www.optimization-online.org/DB_HTML/2019/05/7199.html}
 }
 
 @misc{StructDualDynProg,
   author       = {Benoît Legat},
-  title        = {leopoldcambier/StructDualDynProg.jl: v0.2.0},
+  title        = {JuliaStochOpt/StructDualDynProg.jl: v0.2.0},
   month        = oct,
   year         = 2018,
   url          = {https://github.com/JuliaStochOpt/StructDualDynProg.jl}
@@ -243,4 +245,33 @@ doi = {10.1137/15M1020575},
   author = {Maceira, Maria Elvira Pineiro and Duarte, V. S. and Penna, D. D. J. and Moraes, L. A. M. and Melo, A. C. G.},
   year = {2008},
   pages = {14--18}
+}
+
+@book{bertsekas1995dynamic,
+  title={Dynamic programming and optimal control},
+  author={Bertsekas, Dimitri P and Bertsekas, Dimitri P and Bertsekas, Dimitri P and Bertsekas, Dimitri P},
+  volume={1},
+  number={2},
+  year={1995},
+  publisher={Athena scientific Belmont, MA}
+}
+
+@article{bellman1966dynamic,
+  title={Dynamic programming},
+  author={Bellman, Richard},
+  journal={Science},
+  volume={153},
+  number={3731},
+  pages={34--37},
+  year={1966},
+  publisher={American Association for the Advancement of Science}
+}
+
+@inproceedings{maceiral2018twenty,
+  title={Twenty Years of Application of Stochastic Dual Dynamic Programming in Official and Agent Studies in Brazil-Main Features and Improvements on the NEWAVE Model},
+  author={Maceiral, MEP and Penna, DDJ and Diniz, AL and Pinto, RJ and Melo, ACG and Vasconcellos, CV and Cruz, CB},
+  booktitle={2018 Power Systems Computation Conference (PSCC)},
+  pages={1--7},
+  year={2018},
+  organization={IEEE}
 }
