+++ paper/header.tex
@@ -4,14 +4,15 @@
 
 \author[1, 2]{Andrew W. Rosemberg}
 \author[1, 2]{Alexandre Street}
-\author[1, 2]{Joaquim D. Garcia}
+\author[1, 2, 5]{Joaquim D. Garcia}
 \author[1, 3]{Thuener Silva}
 \author[1, 3]{\\ Davi M. Valladão}
 \author[4]{Oscar Dowson}
 \affil[1]{Laboratory of Applied Mathematical Programming and Statistics (LAMPS)}
-\affil[2]{Department of Eletrical Engineering, PUC-RIO}
+\affil[2]{Department of Electrical Engineering, PUC-RIO}
 \affil[3]{Department of Industrial Engineering, PUC-RIO}
 \affil[4]{Department of Industrial Engineering, Northwestern University, Evanston}
+\affil[5]{PSR-Inc}
 
-\keywords{Julia, Optimization, Energy, Stochastic}
+\keywords{Julia, Optimization, Energy, Stochastic Programming, Optimal Power Flow}
 
