### 2019-07-21 #35 @andrewrosemberg
Hi @frapac and @Lukrosz,
Thank you for your reviews! I just pushed 0a05688489d83478bd03daba1a40e945025fd9c2 changes according to some of the comments of your first reviews.

I will continue working to resolve the remaining points of your comments. Please let me know it you have any questions or additional comments.
