### 2019-07-23 #24 @lccostajr
Hi, unfortunately this message went to Spam folder and I could see it only
now.
I would like to review this paper.

On Mon, 15 Jul 2019 at 06:08, Valentin Churavy <notifications@github.com>
wrote:

> @jd-lara <https://github.com/jd-lara>, @bfpc <https://github.com/bfpc>,
> @farpac <https://github.com/farpac>, @lccostajr
> <https://github.com/lccostajr> would you be willing to review this paper?
>
