### 2019-07-23 f55be6c
change Project.toml docs
