+++ .travis.yml
@@ -28,4 +28,5 @@ after_success:
     # push coverage results to Codecov
     - echo $TRAVIS_JULIA_VERSION
     - julia -e 'import Pkg; cd(Pkg.dir("HydroPowerModels")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
+    - julia --project=docs -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
     - julia --project=docs/ -e 'import Pkg; cd(Pkg.dir("HydroPowerModels")); Pkg.add("Documenter"); include(joinpath("docs", "make.jl"))'
