+++ Project.toml
@@ -8,8 +8,11 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+ECOS = "e2685f51-7e38-5353-a97d-a921fd2c8199"
 Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
+GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
