+++ examples/HydroValleys/case3_soc.jl
@@ -38,7 +38,7 @@ alldata = HydroPowerModels.parse_folder(joinpath(WEAVE_ARGS[:testcases_dir],"cas
 #' Plot power grid graph
 if plot_bool == true
     Random.seed!(seed)
-    plot_grid(alldata[1],node_label=false)
+    HydroPowerModels.plot_grid(alldata[1],node_label=false)
 end
 
 #' Parameters
@@ -62,7 +62,7 @@ end_time = time() - start_time
 
 #' Bounds
 if plot_bool == true
-    plot_bound(m)
+    HydroPowerModels.plot_bound(m)
 end
 
 #' ## Simulation
@@ -82,7 +82,7 @@ using Test
 
 #' ## Plot Aggregated Results
 if plot_bool == true
-    plot_aggregated_results(results)
+    HydroPowerModels.plot_aggregated_results(results)
 end
 
 #' # Annex 1: Case Summary
@@ -92,5 +92,5 @@ end
 
 #' # Annex 2: Plot Results
 if plot_bool == true
-    plotresults(results)
+    HydroPowerModels.plotresults(results)
 end
