+++ src/HydroPowerModels.jl
@@ -18,10 +18,8 @@ include("train.jl")
 include("visualization/visualize_data.jl")
 include("objective.jl")
 
-export  hydrothermaloperation, parse_folder, create_param,
-        plotresults, plotscenarios, set_active_demand!, flat_dict,
-        descriptivestatistics_results, signif_dict, plot_grid, plot_hydro_grid,
-        plot_grid_dispatched, plot_aggregated_results, plot_bound
+export  hydrothermaloperation, create_param,
+        set_active_demand!, flat_dict, signif_dict
 
 Reexport.@reexport using PowerModels, SDDP
 
