+++ .travis.yml
@@ -10,28 +10,6 @@ notifications:
     email: false
 sudo: false
 dist: trusty
-
-# install:
-#   - pip --version
-#   - pip install --user --upgrade pip
-
-# addons:
-#     apt_packages:
-#         - gfortran
-#         - qt5-default
-# script:
-#     - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
-#     #- julia -e 'using Pkg; Pkg.add(PackageSpec(url = "https://github.com/odow/MathOptFormat.jl"))'
-#     #- julia -e 'using Pkg; Pkg.add(PackageSpec(url = "https://github.com/odow/SDDP.jl"))'
-#     - julia -e 'if VERSION >= v"0.7.0-" using Pkg; end; Pkg.test("HydroPowerModels"; coverage=true)'
-# after_success:
-#     # push coverage results to Codecov
-#     - echo $TRAVIS_JULIA_VERSION
-#     - julia -e 'import Pkg; cd(Pkg.dir("HydroPowerModels")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-#     - julia -e 'import Pkg; cd(Pkg.dir("HydroPowerModels")); Pkg.add("Documenter"); include(joinpath("docs", "make.jl"))'
-#     # - julia --project=docs -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
-#     # - julia --project=docs/ -e 'import Pkg; cd(Pkg.dir("HydroPowerModels")); Pkg.add("Documenter"); include(joinpath("docs", "make.jl"))'
-
 codecov: true
 jobs:
   include:
