+++ docs/make.jl
@@ -7,7 +7,7 @@ const EXAMPLES = Any[   "Cases"=>"examples/cases.md"]
 for file in ["case3.jl"]
     filename = joinpath(examples_dir, file)
     md_filename = replace(file, ".jl"=>".md")
-    push!(EXAMPLES,joinpath(docs_dir, md_filename))
+    push!(EXAMPLES,joinpath(docs_dir, "src", md_filename))
     Literate.markdown(filename, joinpath(docs_dir, "src"); documenter=true)
 end
 
