+++ docs/make.jl
@@ -3,11 +3,11 @@ using Documenter, Literate, HydroPowerModels
 examples_dir = joinpath(dirname(dirname(@__FILE__)), "examples/HydroValleys")
 docs_dir = dirname(@__FILE__)
 
-const EXAMPLES = Any[   "Cases"=>"examples/cases.md"]
+const EXAMPLES = Any["examples/cases.md"]
 for file in ["case3.jl"]
     filename = joinpath(examples_dir, file)
     md_filename = replace(file, ".jl"=>".md")
-    push!(EXAMPLES,joinpath(docs_dir, "src", md_filename))
+    push!(EXAMPLES,md_filename)
     Literate.markdown(filename, joinpath(docs_dir, "src"); documenter=true)
 end
 
