+++ docs/src/examples/cases.md
@@ -1,7 +1,5 @@
 Case reports are generated during the package building to test package functionalities and provide examples:
 
-<https://andrewrosemberg.github.io/HydroPowerModels.jl/latest/examples/case3>
-
 <https://andrewrosemberg.github.io/HydroPowerModels.jl/latest/examples/case3_5years>
 
 <https://andrewrosemberg.github.io/HydroPowerModels.jl/latest/examples/case3_ac>
