+++ docs/make.jl
@@ -16,7 +16,7 @@ for file in ["case3.jl"]
     filename = joinpath(examples_dir, file)
     md_filename = replace(file, ".jl"=>".md")
     push!(EXAMPLES,md_filename)
-    Literate.markdown(filename, joinpath(docs_dir, "src"); documenter=true, preprocess = update_testcases_dir)
+    Literate.markdown(filename, joinpath(docs_dir, "src"); documenter=true)
 end
 
 makedocs(
