### 2019-07-28 3ab20f1
update make and travis
