+++ codecov.yml
@@ -1,3 +1,3 @@
-comment: false
+comment: true
 ignore:
   - "./src/visualization"
