### 2019-07-29 11987fc
update visualize func
