+++ docs/make.jl
@@ -5,6 +5,7 @@ docs_dir = dirname(@__FILE__)
 testcases_dir = joinpath(dirname(dirname(@__FILE__)), "testcases")
 
 plot_bool = true
+const EXAMPLES = Any["examples/cases.md"]
 
 makedocs(
     modules = [HydroPowerModels],
