### 2019-07-31 82c0e8d
separate network builder
