+++ .travis.yml
@@ -11,6 +11,11 @@ notifications:
 sudo: false
 dist: trusty
 codecov: true
+
+after_success:
+  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
+#  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())';
+
 jobs:
   include:
     - stage: "Documentation"
