+++ paper/paper.tex
@@ -28,21 +28,18 @@ The hydrothermal dispatch problem is important for the planning and operation of
 
 Solving multistage stochastic programs, however, is a challenging numerical problem. The solution of the problem is intractable and common approximations such as stochastic dynamic programming \cite{bellman1966dynamic} \cite{bertsekas1995dynamic} can suffer from high dimensionality (frequently referred as the curse of dimensionality). One method that partially overcomes the curse of dimensionality is the stochastic dual dynamic programming algorithm of ~\cite{pereira1991multi}.
 
-% <review @Lukrosz>
-% I was taught not to use apostrophes in articles. However, I have seen other articles in different journals where apostrophes are used. Therfore, I encourage you not to use them but the decision is up to you.
-% </review>
 A number of programs implementing SDDP are in-use around the world, ranging from unpublished implementations in academic institutions to professional software such as the product developed by PSR, a software and consulting company, also called \textit{SDDP}\footnote{Note that the ``SDDP'' acronym is used to denote a software product, the original algorithm of \cite{pereira1991multi}, and a more general class of algorithms inspired by the original SDDP algorithm.} \cite{psr_software_2019}. However, until recently, there was no fast, reliable, and open-source implementation of the SDDP algorithm. Without such a tool, researchers and practitioners have not had a common ground for the discussion and analysis of different hydrothermal dispatch formulations and their solutions.
 
 The objective of this work is to build an open source tool, called \href{https://github.com/andrewrosemberg/HydroPowerModels.jl}{\textit{HydroPowerModels.jl}}, that can be this common ground. \textit{HydroPowerModels.jl} can be used to assess the impact of modeling choices during the planning of a hydrothermal power system. These choices include the usage of different network formulations, the consideration of different risk measures, and the planning horizons for uncertain future costs. Addressing these issues provides the research community and the energy industry with a powerful tool for the efficient design of hydrothermal power systems.
 
 To develop a tool that can be used by both researchers and industry professionals, we take advantage of the Julia language~\cite{bezanson2017julia} and two main packages: \href{https://github.com/lanl-ansi/PowerModels.jl}{\textit{PowerModels.jl} ~\cite{8442948}, which implements power flow models for electrical dispatch, \href{https://github.com/odow/SDDP.jl}{\textit{SDDP.jl}} ~\cite{dowson_sddp.jl}, which implements the stochastic dual dynamic programming algorithm}. Both \textit{PowerModels.jl} and \textit{SDDP.jl} handle their respective optimization models through \href{https://github.com/JuliaOpt/JuMP.jl}{\textit{JuMP.jl}} \cite{DunningHuchetteLubin2017}, a Julia package for mathematical optimization. \textit{JuMP.jl} makes it simple to write complex optimization problems and solve them with numerous solvers. \textit{HydroPowerModels.jl} takes advantage from the fact that \textit{PowerModels.jl} and \textit{SDDP.jl} were not only developed in Julia, but also deeply rely on \textit{JuMP.jl} to build and solve mathematical optimization problems. \textit{SDDP.jl} is used to specify the hydro storage dynamics and stochastics of inflows, renewables and loads. \textit{PowerModels.jl} is used to provide multiple network dispatch formulations as a starting point for the \textit{HydroPowerModels.jl} formulation that couples the electrical constraints with hydro constraints and uncertainty.
 
[…]
