### 2019-08-06 #35 @andrewrosemberg
Hi @frapac and @Lukrosz,

Once again, thank you for the thoughtful comments and reviews. I just [pushed](https://github.com/andrewrosemberg/HydroPowerModels.jl/commit/10ecf35357637c6a620f0ebcd7b510d75dcbca91) the modifications I was working on based on your first and final reviews.

The remaining enhancements of the package proposed in the issues and reviews (seminal use-case testing, project.toml enhancement, benchmarks for SDDP libraries, ...) will be the developed from now on!

About the remaining boxes to be checked, do you have additional questions about the implementation, the usage or other points?

I am at your disposal for anything you may need.

Best regards,
Andrew
