### 2019-08-13 6f7406f
fix bug pld and case3
