### 2019-09-04 596caad
add parameter agg plots: nc
