### 2019-09-05 2cfcc65
fix bug plot
