### 2019-09-05 09197d1
resize agg plot
