### 2019-09-10 32e03e6
update reactive plot
