+++ src/visualization/visualize_data.jl
@@ -96,7 +96,7 @@ function plotresults(results::Dict;nc::Int = 3)
 
     # Thermal Reactive Generation
     if results[:params]["model_constructor_grid_forward"] == PowerModels.ACPPowerModel
-        scen_qgen = [[results[:simulations][i][j][:powersystem]["solution"]["gen"]["$gen"]["qg"] for i=1:100, j=1:results[:params]["stages"]]'.*baseMVA for gen =1:ngen]
+        scen_qgen = [[results[:simulations][i][j][:powersystem]["solution"]["gen"]["$gen"]["qg"] for i=1:nsim, j=1:results[:params]["stages"]]'.*baseMVA for gen =1:ngen]
         plt =   [plotscenarios(scen_qgen[gen], title  = "Thermal Reactive Generation $gen",
                 ylabel = "MW",
                 xlabel = "Stages",
@@ -134,7 +134,7 @@ function plotresults(results::Dict;nc::Int = 3)
     # Branch Reactive flow
 
     if results[:params]["model_constructor_grid_forward"] == PowerModels.ACPPowerModel
-        scen_branch_qf = [[results[:simulations][i][j][:powersystem]["solution"]["branch"]["$brc"]["qf"] for i=1:100, j=1:results[:params]["stages"]]'.*baseMVA for brc =1:nbrc]
+        scen_branch_qf = [[results[:simulations][i][j][:powersystem]["solution"]["branch"]["$brc"]["qf"] for i=1:nsim, j=1:results[:params]["stages"]]'.*baseMVA for brc =1:nbrc]
         plt =   [plotscenarios(scen_branch_qf[brc], title  = "Branch Reactive Flow $brc",
                 ylabel = "MW",
                 xlabel = "Stages",
