### 2019-09-12 46bc33e
add discount factor
