+++ src/HydroPowerModels.jl
@@ -33,7 +33,7 @@ Required parameters are:
 -   alldata is a vector of dicts with information of the problem's stages. 
 -   param is a dict containing solution parameters.
 """
-function hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict; build_model::Function=HydroPowerModels.build_opf_powermodels)
+function hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict; build_model::Function=HydroPowerModels.build_opf_powermodels, discount_factor::Float64=1.1)
     # verbose
     if !params["verbose"]
         PowerModels.silence()
@@ -121,7 +121,7 @@ function hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict; buil
         add_deficit_cost(sp, data)
         
         # Stage objective
-        set_objective(sp, data)
+        set_objective(sp, data, t, discount_factor)
 
         # # variable primal start
         # JuMP.MathOptInterface.set.(sp,JuMP.MathOptInterface.VariablePrimalStart(), JuMP.all_variables(sp), NaN)
