### 2019-10-04 #35 @andrewrosemberg
Hi @frapac and @Lukrosz,

Thanks again for the thoughtful comments and reviews.  Do you need any additional information for the remaining points?

Best,
Andrew
