### 2019-10-04 #35 @frapac
Hi @andrewrosemberg . For my part, I think you have addressed all the points that I raised during the review. I am looking forward seeing a clear separation between the optimization pass and the simulation, but this may be a bit premature at the moment. In any case, I would be very interested seeing an open-source package that solves OPF in a multistage stochastic programming setting!
