### 2019-10-04 #35 @andrewrosemberg
Thanks @frapac ! Yes, an pure multistage OPF would be very good, I will look into it!
About the review, I believe all the boxes should be checked in order for it to be over. @vchuravy Please correct me if I am wrong.
