### 2019-10-07 #35 @andrewrosemberg
@frapac and @Lukrosz, thanks for all the work and time you dedicated to this! I hope the package will meet your expectations :)

@vchuravy, thanks for the help in the review process! Please let me know if any additional steps are required.

Best,
Andrew
