### 2019-10-08 #35 @andrewrosemberg
@vchuravy, sorry for the subsequent comment, but occurred to me a question:

During the review process I have changed the article. Do we need to recompile de PDF?
