### 2019-10-08 0b6b08c
fix case3, utilities and visualize
