+++ src/utilities.jl
@@ -107,6 +107,7 @@ function quantile_scen(scen::Array{Float64,2},quants::Array{Float64};output_dict
     if output_dict
         output = Dict()
         for col = 1:length(quants)
+            quant = quants[col]
             output["$(quant*100)%"] = quantiles[:,col]
         end
         return output
@@ -208,7 +209,7 @@ function water_energy_res!(data::Dict,res::Int)
         water_val_spill += water_energy_res!(data,j)
     end
 
-    data["hydro"]["Hydrogenerators"][res]["water_energy"] += (water_val_turn+water_val_spill)/2
+    data["hydro"]["Hydrogenerators"][res]["water_energy"] += water_val_turn >= water_val_spill ? water_val_turn : water_val_spill
     
     return data["hydro"]["Hydrogenerators"][res]["water_energy"]
 end
