### 2019-10-08 6f205b8
change scale to try and avoid numerical issues
