+++ .travis.yml
@@ -12,6 +12,11 @@ sudo: false
 dist: trusty
 codecov: true
 
+script:
+#    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+    - julia -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
+    - julia -e 'Pkg.test("HydroPowerModels"; coverage=true)'
+
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
 #  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())';
