+++ .travis.yml
@@ -14,8 +14,9 @@ codecov: true
 
 script:
 #    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
-    - julia -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
-    - julia -e 'Pkg.test("HydroPowerModels"; coverage=true)'
+    - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
+    - julia --project -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
+    - julia --project -e 'Pkg.test("HydroPowerModels"; coverage=true)'
 
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
