+++ .travis.yml
@@ -13,7 +13,7 @@ dist: trusty
 codecov: true
 
 script:
-#    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
     - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
     - julia --project -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
     - julia --project -e 'Pkg.test("HydroPowerModels"; coverage=true)'
