+++ .travis.yml
@@ -12,11 +12,11 @@ sudo: false
 dist: trusty
 codecov: true
 
-script:
-    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
-    - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
-    - julia --project -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
-    - julia --project -e 'Pkg.test("HydroPowerModels"; coverage=true)'
+# script:
+#     - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+#     - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
+#     - julia --project -e 'Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
+#     - julia --project -e 'Pkg.test("HydroPowerModels"; coverage=true)'
 
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
