+++ .travis.yml
@@ -19,6 +19,7 @@ codecov: true
 #     - julia --project -e 'Pkg.test("HydroPowerModels"; coverage=true)'
 
 after_success:
+  - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
 #  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())';
 
