### 2019-10-09 c355b76
update travis and docs
