+++ .travis.yml
@@ -13,26 +13,35 @@ notifications:
 branches:
   only:
     - master
+stages:
+  - name: test
+  - name: docs
 codecov: true
 
 # codecov: true
 # coveralls: true
 
 script:
-    # - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+    - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
     - julia --project -e 'using Pkg; Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
-    - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
-
-# after_success:
-#   - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
-#   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
-# #  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())';
+    - julia --project -e 'using Pkg; Pkg.build(); Pkg.test("HydroPowerModels"; coverage=true)'
 
 jobs:
   include:
-    - stage: "Documentation"
+    - stage: test
+      julia: 1.0
+      os: linux
+      script:
+        - julia --project -e 'using Pkg; Pkg.clone(pwd()); Pkg.build("HydroPowerModels")'
+        - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
+      after_success: skip
+    - stage: docs
       julia: 1.0
       os: linux
       script:
         - julia --project=docs -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
         - julia --project=docs --color=yes docs/make.jl
+
+after_success:
+  - julia --project=test/coverage -e 'using Pkg; Pkg.instantiate();
+        using Coverage; Codecov.submit(Codecov.process_folder())'
