+++ examples/HydroValleys/case3.jl
@@ -72,7 +72,7 @@ results
 # ## Testing Results
 using Test
 # Bound
-@test isapprox(SDDP.calculate_bound(m.policygraph), 11298.218756240378, atol=1)
+@test isapprox(SDDP.calculate_bound(m.policygraph), 11298.218756240378, atol=2)
 # Number of Simulations
 @test length(results[:simulations]) == 100
 
