+++ examples/HydroValleys/case3_5years.jl
@@ -53,7 +53,7 @@ m = hydrothermaloperation(alldata, params);
 #' ## Train
 #+ results =  "hidden"
 start_time = time()
-HydroPowerModels.train(m,iteration_limit = 100,stopping_rules= [SDDP.Statistical(num_replications = 20,iteration_period=20)]);
+HydroPowerModels.train(m,iteration_limit = 100,stopping_rules= [SDDP.Statistical(num_replications = 200,iteration_period=80)]);
 end_time = time() - start_time
 
 #' Termination Status and solve time (s)
@@ -73,7 +73,7 @@ results
 #' ## Testing Results
 using Test
 #' Bound
-@test isapprox(SDDP.calculate_bound(m.policygraph), 76487.72, atol=1)
+@test isapprox(SDDP.calculate_bound(m.policygraph), 59357.12, atol=10)
 #' Number of Simulations
 @test length(results[:simulations]) == 100
 
