+++ Project.toml
@@ -39,5 +39,5 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Reel = "71555da5-176e-5e73-a222-aebc6c6e4f2f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
-[targets]	
+[targets]
 test = ["Test", "GLPK", "ECOS"]
