+++ src/variable.jl
@@ -28,7 +28,7 @@ function variable_volume(sp, data::Dict)
 end
 
 """creates deficit variables"""
-function variable_deficit(sp, data::Dict, pm::GenericPowerModel)
+function variable_deficit(sp, data::Dict, pm::AbstractPowerModel)
     PowerModels.var(pm, 0, 1)[:deficit] = @variable(sp, deficit[i in collect(1:length(PowerModels.con(pm, 0, 1, :kcl_p)))] >= 0)
 end
 
