+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"examples/case3_cmp_formulations/#Example-Case-3-Comparing-Formulations-Year-Planning-1","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"","category":"section"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"author : Andrew Rosemberg ","category":"page"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"date : 21th Feb 2019 ","category":"page"},{"location":"examples/case3_cmp_formulations/#Introduction-1","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Introduction","text":"","category":"section"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"This an example of the HydroPowerModels package for solving a simple stochastic case with the following specifications: ","category":"page"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"- 3 Buses - 3 Lines - 2 Generators - 1 Reservoir and Hydrogenerator \n\n- 3 Scenarios - 12 Stages \n\n- DC,SOC and AC Formulations","category":"page"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"Simulation's results will be shown in tables to facilitate comparison. Each formulation will have its own table where the columns will be named according to the variables of the problem: ","category":"page"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"- \"[variable type]_[index]_[quantile]\".","category":"page"},{"location":"examples/case3_cmp_formulations/#","page":"Example Case 3 Comparing Formulations - Year Planning","title":"Example Case 3 Comparing Formulations - Year Planning","text":"Variable types in this example:","category":"p
[…]
