### 2019-10-16 cf95c60
fix plots patch and soc case 3
