### 2019-10-16 813571d
try avoid numerical issue
