+++ examples/HydroValleys/case3_soc.jl
@@ -28,7 +28,7 @@ if !@isdefined plot_bool
     plot_bool = true
 end
 using Random
-seed = 1111
+seed = 1221
 
 #' ## Load Case Specifications
 
