### 2019-10-17 850d6ec
refactor discount factor
