### 2019-10-18 951e77e
separate graph from policy graph creation
