### 2019-10-19 f873f13
compat simulate with discount factors
