### 2019-10-22 92b524b
add mean to quantile dict
