### 2019-11-14 cbdb6a6
add reactive cost
