+++ src/visualization/visualize_data.jl
@@ -1,6 +1,6 @@
 using Plots, Plots.PlotMeasures
-import Cairo, Fontconfig
-using LightGraphs, GraphPlot, Compose
+# import Cairo, Fontconfig
+# using LightGraphs, GraphPlot, Compose
 using Random, Reel
 
 """
@@ -69,234 +69,234 @@ Parameters:
 -   nc             : Number of figures per line.
 
 """
-function plotresults(results::Dict;nc::Int = 3)
-    plt_total = Array{Plots.Plot}(undef,10000)
-    nplots = 0
-    nsim = length(results[:simulations])
-    nstages = length(results[:simulations][1])
+# function plotresults(results::Dict;nc::Int = 3)
+#     plt_total = Array{Plots.Plot}(undef,10000)
+#     nplots = 0
+#     nsim = length(results[:simulations])
+#     nstages = length(results[:simulations][1])
 
-    # Thermal Generation
-    ngen = length(results[:data][1]["powersystem"]["gen"])
-    idxhyd = idx_hydro(results[:data][1])
-    idxgen = setdiff(collect(1:ngen),idxhyd)
-    baseMVA =  [results[:simulations][i][j][:powersystem]["solution"]["baseMVA"] for i=1:nsim, j=1:nstages]'
-    scen_gen = [[results[:simulations][i][j][:powersystem]["solution"]["gen"]["$gen"]["pg"] for i=1:nsim, j=1:nstages]'.*baseMVA for gen =1:ngen]
+#     # Thermal Generation
+#     ngen = length(results[:data][1]["powersystem"]["gen"])
+#     idxhyd = idx_hydro(results[:data][1])
+#     idxgen = setdiff(collect(1:ngen),idxhyd)
+#     baseMVA =  [results[:simulations][i][j][:powersystem]["solution"]["baseMVA"] for i=1:nsim, j=1:nstages]'
+#     scen_gen = [[results[:simulations][i][j][:powersystem]["solution"]["gen"]["$gen"]["pg"] for i=1:nsim, j=1:nstages]'.*baseMVA for gen =1:ngen]
 
-    plt =   [plotscenarios(scen_gen[gen], title  = "Thermal Generation $gen",
-                ylabel = "MW",
-                xlabel = "Stages",
-                bottom_margin = 10mm,
-                right_margin = 10mm,
-                left_margin = 10mm                
-                )
-            for gen in idxgen
-    ]
-    plt_total[nplots+1:nplots+length(plt)] = plt
-    nplots +=length(plt) 
+#     plt =   [plotscenarios(scen_gen[gen], title  = "Thermal Generation $gen",
+#                 ylabel = "MW",
+#                 xlabel = "Stages",
+#                 bottom_margin = 10mm,
+#                 right_margin = 10mm,
+#                 left_margin = 10mm                
+#                 )
+#             for gen in idxgen
+#     ]
+#     plt_total[nplots+1:nplots+length(plt)] = plt
+#     nplots +=length(plt) 
 
-    # Thermal Reactive Generation
-    if results[:params]["model_constructor_grid_forward"] == PowerModels.ACPPowerModel
-        scen_qgen = [[results[:simulations][i][j][:powersystem]["solution"]["gen"]["$gen"]["qg"] for i=1:nsim, j=1:results[:params]["stages"]]'.*baseMVA for gen =1:ngen]
-        plt =   [plotscenarios(scen_qgen[gen], title  = "Thermal Reactive Generation $gen",
-                ylabel = "MW",
-                xlabel = "Stages",
-                bottom_margin = 10mm,
-                right_margin = 10mm,
-                left_margin = 10mm                
-                )
-            for gen in idxgen
-        ]
-        plt_total[nplots+1:nplots+length(plt)] = plt
-        nplots +=length(plt) 
-    end
[…]
