### 2020-05-24 #35 @andrewrosemberg
@vchuravy, I have published the code using zenodo, following the proceedings [docs](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance), and the DOI associated with the final version of the paper and package is: `10.5281/zenodo.3842130`
The version is `v0.1.0`.
Doi is shown in the [master](https://github.com/andrewrosemberg/HydroPowerModels.jl/tree/master).
Tag is [this one](https://github.com/andrewrosemberg/HydroPowerModels.jl/tree/v0.1.0).
