### 2020-07-28 bb13ac5
add missing DOIs to ref.bib
