+++ paper/ref.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi = {10.1137/141000671}
 }
 
 @article{Moore1990,
@@ -42,7 +43,8 @@
   number={3},
   pages={161--169},
   year={1989},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/0142-0615(89)90025-2}
 }
 
 @article{pereira1991multi,
@@ -53,7 +55,8 @@
   number={1-3},
   pages={359--375},
   year={1991},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/bf01582895}
 }
 
 @article{philpott2013solving,
@@ -64,7 +67,8 @@
   number={4},
   pages={957--970},
   year={2013},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi = {10.1287/opre.2013.1175}
 }
 
 @article{baucke2017deterministic,
@@ -195,7 +199,8 @@ doi = {10.1137/15M1020575},
   number={1-2},
   pages={1--221},
   year={2019},
-  publisher={Now Publishers, Inc.}
+  publisher={Now Publishers, Inc.},
+  doi = {10.1561/9781680835410}
 }
 
 @article{donohue2006abridged,
@@ -215,7 +220,8 @@ doi = {10.1137/15M1020575},
   number={3},
   pages={497--524},
   year={1999},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1023/a:1022641805263}
 }
 
 @misc{psr_software_2019,
@@ -273,5 +279,6 @@ doi = {10.1137/15M1020575},
   booktitle={2018 Power Systems Computation Conference (PSCC)},
   pages={1--7},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.23919/pscc.2018.8442754}
 }
