### 2020-07-28 #35 @andrewrosemberg
@vchuravy Missing DOIs added! sorry for the delay.
The updated version is already in the master, do I need to make a new release?
