### 2020-07-29 5d928d3
remove gif dependencies
