### 2020-07-29 0cc8513
update julia travis and remove gif deps from extra
