+++ .travis.yml
@@ -2,7 +2,7 @@ language: julia
 os:
     - linux
 julia:
-    - 1.0
+    - 1.3
 notifications:
     email: false
 branches:
@@ -18,14 +18,14 @@ script:
 jobs:
   include:
     - stage: test
-      julia: 1.0
+      julia: 1.3
       os: linux
       script:
         - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
       after_success:
         - julia --project -e 'import Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
     - stage: docs
-      julia: 1.0
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
