### 2020-07-29 b79ce7a
fix plots on cases
