+++ examples/HydroValleys/case3_soc.jl
@@ -45,7 +45,7 @@ end
 params = create_param(  stages = 12, 
                         model_constructor_grid  = SOCWRConicPowerModel,
                         post_method             = PowerModels.post_opf,
-                        optimizer               = with_optimizer(COSMO.Optimizer, atol=1e-10, rtol=1e-10))
+                        optimizer               = with_optimizer(COSMO.Optimizer))
 
 #' ## Build Model
 #+ results =  "hidden"
