### 2020-07-29 451ad38
fix plots examples docs
