### 2020-07-29 7649843
Ar/update deps docs (#39)
* rm wrong solver params

* update toml

* fix plots examples docs

Co-authored-by: Andrew.Rosemberg <andrew.rosemberg@invenialabs.co.uk>
