### 2020-10-26 3854d4b
update travis more tests
