+++ .travis.yml
@@ -8,6 +8,8 @@ julia:
     - nightly
 notifications:
     email: false
+  allow_failures:
+    - julia: nightly
 branches:
   only:
     - master
