+++ .travis.yml
@@ -8,7 +8,7 @@ julia:
     - nightly
 notifications:
     email: false
-  allow_failures:
+allow_failures:
     - julia: nightly
 branches:
   only:
@@ -23,6 +23,8 @@ script:
 jobs:
   include:
     - stage: test
+      allow_failures:
+        - julia: nightly
       script:
         - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
       after_success:
