+++ .travis.yml
@@ -6,10 +6,11 @@ julia:
     - 1.3
     - 1
     - nightly
+matrix:
+  allow_failures:
+          - julia: nightly
 notifications:
     email: false
-allow_failures:
-    - julia: nightly
 branches:
   only:
     - master
@@ -23,8 +24,6 @@ script:
 jobs:
   include:
     - stage: test
-      allow_failures:
-        - julia: nightly
       script:
         - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
       after_success:
