+++ .travis.yml
@@ -6,9 +6,6 @@ julia:
     - 1.3
     - 1
     - nightly
-matrix:
-  allow_failures:
-          - julia: nightly
 notifications:
     email: false
 branches:
@@ -22,6 +19,8 @@ script:
     - julia --project -e 'using Pkg; Pkg.build(); Pkg.test("HydroPowerModels"; coverage=true)'
 
 jobs:
+  allow_failures:
+    - julia: nightly
   include:
     - stage: test
       script:
