### 2020-10-26 9cd1d9d
Add tests for other OS and julia versions (#41)
* add new tests

* allow failure on nightly

* allow_failures jobs nightly

* fix travis

* fix travis

* update compat

* update julia compat

* update docs manifest
