+++ .travis.yml
@@ -1,8 +1,11 @@
 language: julia
 os:
     - linux
+    - osx
 julia:
     - 1.3
+    - 1
+    - nightly
 notifications:
     email: false
 branches:
@@ -16,10 +19,10 @@ script:
     - julia --project -e 'using Pkg; Pkg.build(); Pkg.test("HydroPowerModels"; coverage=true)'
 
 jobs:
+  allow_failures:
+    - julia: nightly
   include:
     - stage: test
-      julia: 1.3
-      os: linux
       script:
         - julia --project -e 'using Pkg; Pkg.test("HydroPowerModels"; coverage=true)'
       after_success:
