+++ Project.toml
@@ -1,7 +1,7 @@
 name = "HydroPowerModels"
 uuid = "baadc3fc-4b32-11e9-110e-59741390feb0"
 authors = ["andrewrosemberg <andrew@ele.puc-rio.br>"]
-version = "0.1.0"
+version = "0.1.1"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
