### 2020-10-27 f3dcdab
update installation
